codebases/webapp_trials/flask_alpaca/binance_testnet/requirements.txt
python-binance
TA-Lib
numpy
websocket_client
pprintpp